Move binary syntax combinator to punctuation combinator
Before this change, it was raising a `FunctionClauseError`
in the post process function `Makeup.Lexers.ErlangLexer.__as_erlang_language/1`
when trying to lex text with binary syntax in it (`<<` and `>>`).

It was erroring out because we were not wrapping the binary syntax
combinators with the `Makeup.Lexer.Combinators.token/2` combinator,
which returns the result that `__as_erlang_language/1` expects.

It was passing a literal `"<<"` instead of the `token` tuple that is expected.
e.g. `{:token_name, attrs = %{}, literal = "<<"}`.